FIX: Rename branding to logo and fonts
Change branding page into logo and fonts.
