--- app/assets/javascripts/admin/addon/routes/admin-config-branding.js
@@ -1,8 +1,0 @@
-import DiscourseRoute from "discourse/routes/discourse";
-import { i18n } from "discourse-i18n";
-
-export default class AdminConfigBrandingRoute extends DiscourseRoute {
-  titleToken() {
-    return i18n("admin.config.branding.title");
-  }
-}
