+++ app/assets/javascripts/admin/addon/templates/config-logo-and-fonts.gjs
@@ -2,34 +2,34 @@ import RouteTemplate from "ember-route-template";
 import DBreadcrumbsItem from "discourse/components/d-breadcrumbs-item";
 import DPageHeader from "discourse/components/d-page-header";
 import { i18n } from "discourse-i18n";
-import AdminBrandingFontsForm from "admin/components/admin-branding-fonts-form";
-import AdminBrandingLogoForm from "admin/components/admin-branding-logo-form";
 import AdminConfigAreaCard from "admin/components/admin-config-area-card";
+import AdminFontsForm from "admin/components/admin-fonts-form";
+import AdminLogoForm from "admin/components/admin-logo-form";
 
 export default RouteTemplate(
   <template>
     <div class="admin-config-page">
       <DPageHeader
         @hideTabs={{true}}
-        @titleLabel={{i18n "admin.config.branding.title"}}
+        @titleLabel={{i18n "admin.config.logo_and_fonts.title"}}
       >
         <:breadcrumbs>
           <DBreadcrumbsItem @path="/admin" @label={{i18n "admin_title"}} />
           <DBreadcrumbsItem
-            @path="/admin/config/branding"
-            @label={{i18n "admin.config.branding.title"}}
+            @path="/admin/config/logo-and-fonts"
+            @label={{i18n "admin.config.logo_and_fonts.title"}}
           />
         </:breadcrumbs>
       </DPageHeader>
       <div class="admin-config-area">
         <div class="admin-config-area__primary-content">
           <AdminConfigAreaCard
-            @heading="admin.config.branding.logo.title"
+            @heading="admin.config.logo_and_fonts.logo.title"
             @collapsable={{true}}
-            class="admin-config-area-branding__logo"
+            class="admin-config-area-logo-and-fonts__logo"
           >
             <:content>
-              <AdminBrandingLogoForm />
+              <AdminLogoForm />
             </:content>
           </AdminConfigAreaCard>
         </div>
@@ -37,12 +37,12 @@ export default RouteTemplate(
       <div class="admin-config-area">
         <div class="admin-config-area__primary-content">
           <AdminConfigAreaCard
-            @heading="admin.config.branding.fonts.title"
+            @heading="admin.config.logo_and_fonts.fonts.title"
             @collapsable={{true}}
-            class="admin-config-area-branding__fonts"
+            class="admin-config-area-logo-and-fonts__fonts"
           >
             <:content>
-              <AdminBrandingFontsForm />
+              <AdminFontsForm />
             </:content>
           </AdminConfigAreaCard>
         </div>
