+++ app/assets/stylesheets/admin/admin_config_color_palettes.scss
@@ -14,7 +14,7 @@
       margin-bottom: 1em;
     }
 
-    &__fonts-and-logos-hint {
+    &__logo-and-fonts-hint {
       display: flex;
       justify-content: space-between;
     }
