+++ config/locales/client.en.yml
@@ -5256,8 +5256,8 @@ en:
         login_and_authentication:
           title: "Login & authentication"
           header_description: "Configure how users log in and authenticate, secrets and keys, OAuth2 providers, and more"
-        branding:
-          title: "Branding"
+        logo_and_fonts:
+          title: "Logo & fonts"
           header_description: "Customize the variations of your site logo and fonts"
           logo:
             title: "Logo"
@@ -6275,8 +6275,8 @@ en:
             toggle_description: "Color palette can be selected by users"
           colors:
             title: "Colors"
-          fonts_and_logos_hint: "Looking to update your fonts and logos?"
-          go_to_branding: "Go to Branding"
+          logo_and_fonts_hint: "Looking to update your logo and fonts?"
+          go_to_logo_and_fonts: "Go to Logo & fonts"
           save_changes: "Save changes"
           unsaved_changes: "You have unsaved changes"
           copy_of: "Copy of %{name}"
