+++ spec/system/page_objects/components/admin_logo_form.rb
@@ -2,7 +2,7 @@
 
 module PageObjects
   module Components
-    class AdminBrandingLogoForm < PageObjects::Components::Base
+    class AdminLogoForm < PageObjects::Components::Base
       def image_uploader(image_type)
         PageObjects::Components::UppyImageUploader.new(
           find(".form-kit__container[data-name='#{image_type}']"),
@@ -49,7 +49,7 @@ module PageObjects
 
       def has_saved_successfully?
         PageObjects::Components::Toasts.new.has_success?(
-          I18n.t("admin_js.admin.config.branding.logo.form.saved"),
+          I18n.t("admin_js.admin.config.logo_and_fonts.logo.form.saved"),
         )
       end
